Allow the frontend lookup keys to be specified by an optional function in the config file
frontendLookup(hostKey,config) is an optional function to dynamically
override the format of the Redis keys used to lookup the backends. If
frontendLookup is not defined, Hipache will do three lookups for each
incoming request: one matching exactly the host header of the HTTP
request, another using a '*' + the last two components of this host
header (in order to process wildcards) and a final one to match frontend:*
to match a global wildcard.

If frontendLookup is defined, then the lookup is done using the array
of keys returned by calling frontendLookup. Note: the frontend: is kept
at all time.

Remember that this function, if specified, will be called for every page
request so it is important that it is not slow.

Here is an example of using only the machine name as the lookup key

  "frontendLookup": (function( hostKey, config ){
    var os = require('os');
    var key = os.hostname().toLowerCase();
    var frontends = [key];

    return function(){ return frontends };
  }())